Backend/requirements.txt
fastapi==0.120.4
httpx==0.27.0
uvicorn==0.38.0
python-multipart==0.0.20

# ✅ LangChain + Google GenAI stack (v1 API)
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
google-generativeai==0.8.3

# ✅ Vector + Database
chromadb==1.3.0
sqlalchemy==2.0.23
aiosqlite==0.20.0
aiofiles==23.2.1

# ✅ Auth & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# ✅ PDF, Reporting & Utilities
reportlab==4.4.4
pypdf==6.1.3
python-dotenv==1.0.1
pydantic==2.9.2

# ✅ Optional helper (recommended)
google-auth==2.35.0          # Helps with ADC setup locally
google-auth-oauthlib==1.2.1  # For token-based auth (if needed)
